Move public util functions to util/status.go
